메인화면에서 채널 등록하는 코드 수정
