+++ package.json
@@ -12,7 +12,6 @@
     "author": "Owake",
     "license": "ISC",
     "dependencies": {
-        "body-parser": "^1.19.0",
         "cross-env": "^7.0.3",
         "dotenv": "^10.0.0",
         "ejs": "^3.1.6",
