+++ public/utils/common.js
@@ -7,7 +7,6 @@ import { momentShare } from "./parts/momentShare.js";
 import { muteOptions } from "./parts/muteUtils.js";
 import { optionsBtnFunc } from "./parts/options-btn.js";
 import { deviceSettings } from "./parts/recodingDeviceCtrl.js";
-import { channelCreate } from "./parts/channelCreate.js";
 
 screenShareFunc();
 remoteDisplay();
@@ -18,4 +17,3 @@ momentShare();
 muteOptions();
 optionsBtnFunc();
 deviceSettings();
-channelCreate();
