+++ views/index.ejs
@@ -49,8 +49,7 @@
         </div>
     </div>
 
-    <script src="rtcClient.js"></script>
-    <script src="common.js" type="module"></script>
+    <script src="channelCreate.js"></script>
     <script>
         const bodyContainer = document.querySelector(".body-container");
         const channelBox = document.querySelector(".channel-box");
